requirements.txt
flask>=3.0.0
google-genai>=1.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
werkzeug>=3.0.0
